requirements.txt
pyinstaller>=6.0.0
colorama>=0.4.6
Pillow>=10.0.0
